+++ app/api/deps.py
@@ -1,11 +1,10 @@
 from typing import Generator
-from fastapi.security import OAuth2PasswordBearer
 
 import redis
+from fastapi.security import OAuth2PasswordBearer
 
 from app.core.config import settings
 from app.database.session import SessionLocal
-
 
 reusable_oauth2 = OAuth2PasswordBearer(tokenUrl=f"/api/v1/login/access-token")
