+++ app/core/logger.py
@@ -3,7 +3,6 @@ Logging configuration to be set for the server
 """
 from pydantic import BaseModel
 
-
 from app.core.config import settings
