+++ app/core/security.py
@@ -3,14 +3,14 @@ from datetime import datetime, timedelta
 from typing import Any, Optional, Tuple, Union
 from urllib.parse import urljoin
 from uuid import uuid4
-from fastapi import Depends, Request
 
 import httpx
 import jwt
 from cryptography.fernet import Fernet
+from fastapi import Depends, Request
 from passlib.context import CryptContext
-from app.api.deps import get_redis_client
 
+from app.api.deps import get_redis_client
 from app.common_tags import ONADATA_TOKEN_ENDPOINT
 from app.core.config import settings
 from app.models.server import Server
