+++ app/crud/__init__.py
@@ -1,4 +1,4 @@
-from .crud_server import server
-from .crud_user import user
 from .crud_configuration import configuration
 from .crud_hyperfile import hyperfile
+from .crud_server import server
+from .crud_user import user
